refactor: Move `add_torrent` method from `EngineActor` to `Engine`
